@@ -1,5 +1,5 @@
 buildscript {
-    ext.kotlin_version = '1.2.10'
+    ext.kotlin_version = '1.3.21'
 
     repositories {
         mavenCentral()
@@ -23,6 +23,10 @@ repositories {
 
 dependencies {
     compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    compile 'com.google.guava:guava:16.0'
-    testCompile 'junit:junit:4.12'
+    compile 'com.google.guava:guava:28.0-jre'
+    testCompile 'org.junit.jupiter:junit-jupiter:5.4.0'
+}
+
+test {
+    useJUnitPlatform()
 }
