@@ -1,10 +1,11 @@
 package i_introduction._0_Hello_World
 
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N00StartKtTest {
-    @Test fun testOk() {
+    @Test
+    fun testOk() {
         assertEquals("OK", task0())
     }
 }
