@@ -1,10 +1,11 @@
 package i_introduction._10_Object_Expressions
 
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N10ObjectExpressionsKtTest {
-    @Test fun testSort() {
+    @Test
+    fun testSort() {
         assertEquals(listOf(5, 2, 1), task10())
     }
 }
