@@ -1,10 +1,11 @@
 package i_introduction._11_SAM_Conversions
 
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N11SAMConversionsKtTest {
-    @Test fun testSort() {
+    @Test
+    fun testSort() {
         assertEquals(listOf(5, 2, 1), task11())
     }
 }
