@@ -1,10 +1,12 @@
 package i_introduction._2_Named_Arguments
 
-import org.junit.Assert.assertEquals
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N02NamedArgumentsKtTest {
 
-    @org.junit.Test fun testJoinToString() {
+    @Test
+    fun testJoinToString() {
         assertEquals("{1, 2, 3, 42, 555}", task2(listOf(1, 2, 3, 42, 555)))
     }
 
