@@ -1,11 +1,12 @@
 package i_introduction._3_Default_Arguments
 
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N03DefaultArgumentsKtTest {
 
-    @Test fun testDefaultAndNamedParams() {
+    @Test
+    fun testDefaultAndNamedParams() {
         assertEquals("a42b1C42D2", task3())
     }
 }
