@@ -1,14 +1,16 @@
 package i_introduction._9_Extension_Functions
 
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N09ExtensionFunctionsKtTest {
-    @Test fun testIntExtension() {
-        assertEquals("Rational number creation error: ", RationalNumber(4, 1), 4.r())
+    @Test
+    fun testIntExtension() {
+        assertEquals(RationalNumber(4, 1), 4.r(), "Rational number creation error: ")
     }
 
-    @Test fun testPairExtension() {
-        assertEquals("Rational number creation error: ", RationalNumber(2, 3), Pair(2, 3).r())
+    @Test
+    fun testPairExtension() {
+        assertEquals(RationalNumber(2, 3), Pair(2, 3).r(), "Rational number creation error: ")
     }
 }
