@@ -2,11 +2,12 @@ package ii_collections
 
 import ii_collections.data.customers
 import ii_collections.data.shop
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N13IntroductionKtTest {
-    @Test fun testSetOfCustomers() {
+    @Test
+    fun testSetOfCustomers() {
         assertEquals(customers.values.toSet(), shop.getSetOfCustomers())
     }
 }
