@@ -1,18 +1,20 @@
 package ii_collections
 
 import ii_collections.data.*
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N14FilterMapKtTest {
-    @Test fun testCitiesCustomersAreFrom() {
+    @Test
+    fun testCitiesCustomersAreFrom() {
         assertEquals(setOf(Canberra, Vancouver, Budapest, Ankara, Tokyo), shop.getCitiesCustomersAreFrom())
     }
 
     /**
      * Returns the list of the customers who live in the city 'city'
      */
-    @Test fun testCustomersFromCity() {
+    @Test
+    fun testCustomersFromCity() {
         assertEquals(listOf(customers[lucas], customers[cooper]), shop.getCustomersFrom(Canberra))
     }
 }
