@@ -2,11 +2,12 @@ package ii_collections
 
 import ii_collections.data.shop
 import ii_collections.data.sortedCustomers
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N18SortKtTest {
-    @Test fun testGetCustomersSortedByNumberOfOrders() {
+    @Test
+    fun testGetCustomersSortedByNumberOfOrders() {
         assertEquals(sortedCustomers, shop.getCustomersSortedByNumberOfOrders())
     }
 }
