@@ -1,31 +1,35 @@
 package ii_collections
 
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N24ExtensionsOnCollectionsKtTest {
-    @Test fun testCollectionOfOneElement() {
+    @Test
+    fun testCollectionOfOneElement() {
         doTest(listOf("a"), listOf("a"))
     }
 
-    @Test fun testEmptyCollection() {
+    @Test
+    fun testEmptyCollection() {
         doTest(null, listOf())
     }
 
-    @Test fun testSimpleCollection() {
+    @Test
+    fun testSimpleCollection() {
         doTest(listOf("a", "c"), listOf("a", "bb", "c"))
     }
 
-    @Test fun testCollectionWithEmptyStrings() {
+    @Test
+    fun testCollectionWithEmptyStrings() {
         doTest(listOf("", "", "", ""), listOf("", "", "", "", "a", "bb", "ccc", "dddd"))
     }
 
-    @Test fun testCollectionWithTwoGroupsOfMaximalSize() {
+    @Test
+    fun testCollectionWithTwoGroupsOfMaximalSize() {
         doTest(listOf("a", "c"), listOf("a", "bb", "c", "dd"))
     }
 
     private fun doTest(expected: Collection<String>?, argument: Collection<String>) {
-        assertEquals("The function 'doSomethingStrangeWithCollection' should do at least something with a collection:",
-                expected, doSomethingStrangeWithCollection(argument))
+        assertEquals(expected, doSomethingStrangeWithCollection(argument), "The function 'doSomethingStrangeWithCollection' should do at least something with a collection:")
     }
 }
