@@ -1,18 +1,18 @@
 package iv_properties
 
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N32PropertiesKtTest {
-    @Test fun testPropertyWithCounter() {
+    @Test
+    fun testPropertyWithCounter() {
         val q = PropertyExample()
         q.propertyWithCounter = 14
         q.propertyWithCounter = 21
         q.propertyWithCounter = 32
-        assertEquals("The property q.counter should contain the number of assignments to q.propertyWithCounter:",
-                3, q.counter)
+        assertEquals(3, q.counter, "The property q.counter should contain the number of assignments to q.propertyWithCounter:")
         // Here we have to use !! due to false smart cast impossible
-        assertEquals("The property q.propertyWithCounter should be set:", 32, q.propertyWithCounter!!)
+        assertEquals(32, q.propertyWithCounter!!, "The property q.propertyWithCounter should be set:")
     }
 
 }
