@@ -6,28 +6,29 @@ import i_introduction._4_Lambdas.N04LambdasKtTest
 import i_introduction._7_Nullable_Types.N07NullableTypesKtTest
 import i_introduction._8_Smart_Casts.N08SmartCastsKtTest
 import ii_collections.N24ExtensionsOnCollectionsKtTest
-import org.junit.Assert.assertTrue
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertTrue
+import org.junit.jupiter.api.Test
 
 class AdditionalTest {
-    @Test fun test() {
+    @Test
+    fun test() {
         invokeTests(
-                { N01JavaToKotlinConverterKtTest().collection() },
-                { N03DefaultArgumentsKtTest().testDefaultAndNamedParams() },
-                { N04LambdasKtTest().contains() },
-                { N07NullableTypesKtTest().everythingIsOk() },
-                { N08SmartCastsKtTest().testNum() },
-                { N24ExtensionsOnCollectionsKtTest().testCollectionOfOneElement() }
+            { N01JavaToKotlinConverterKtTest().collection() },
+            { N03DefaultArgumentsKtTest().testDefaultAndNamedParams() },
+            { N04LambdasKtTest().contains() },
+            { N07NullableTypesKtTest().everythingIsOk() },
+            { N08SmartCastsKtTest().testNum() },
+            { N24ExtensionsOnCollectionsKtTest().testCollectionOfOneElement() }
         )
-        assertTrue("${JavaCode.set}", JavaCode.set.isEmpty())
+        assertTrue(JavaCode.set.isEmpty(), "${JavaCode.set}")
     }
 
     private fun invokeTests(vararg tests: () -> Unit) {
         for (test in tests) {
             try {
                 test()
+            } catch (e: NotImplementedException) {
             }
-            catch (e: NotImplementedException) {}
         }
     }
 }
