@@ -1,13 +1,13 @@
 package v_builders
 
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class N36ExtensionFunctionLiteralsKtTest {
-    @Test fun testIsOddAndIsEven() {
+    @Test
+    fun testIsOddAndIsEven() {
         val result = task36()
-        assertEquals("The functions 'isOdd' and 'isEven' should be implemented correctly",
-                listOf(false, true, true), result)
+        assertEquals(listOf(false, true, true), result, "The functions 'isOdd' and 'isEven' should be implemented correctly")
 
     }
 }
