@@ -1,11 +1,12 @@
 package v_builders.examples
 
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 import java.util.*
 
 class N38TheFunctionApplyKtTest {
-    @Test fun testBuildString() {
+    @Test
+    fun testBuildString() {
         val expected = StringBuilder().apply {
             append("Numbers: ")
             for (i in 1..10) {
@@ -13,10 +14,11 @@ class N38TheFunctionApplyKtTest {
             }
         }.toString()
         val actual = buildString()
-        assertEquals("String should be built:", expected, actual)
+        assertEquals(expected, actual, "String should be built:")
     }
 
-    @Test fun testBuildMap() {
+    @Test
+    fun testBuildMap() {
         val expected = HashMap<Int, String>().apply {
             put(0, "0")
             for (i in 1..10) {
@@ -24,6 +26,6 @@ class N38TheFunctionApplyKtTest {
             }
         }
         val actual = buildMap()
-        assertEquals("Map should be filled with the right values", expected, actual)
+        assertEquals(expected, actual, "Map should be filled with the right values")
     }
 }
