@@ -23,13 +23,13 @@ fun todoTask7(client: Client?, message: String?, mailer: Mailer): Nothing = TODO
 )
 
 fun sendMessageToClient(
-        client: Client?, message: String?, mailer: Mailer
+    client: Client?, message: String?, mailer: Mailer
 ) {
     todoTask7(client, message, mailer)
 }
 
-class Client (val personalInfo: PersonalInfo?)
-class PersonalInfo (val email: String?)
+class Client(val personalInfo: PersonalInfo?)
+class PersonalInfo(val email: String?)
 
 interface Mailer {
     fun sendMessage(email: String, message: String)
