@@ -7,12 +7,14 @@ fun example() {
     result == listOf('a', 'b', 'c', '1', '2')
 }
 
-val Customer.orderedProducts: Set<Product> get() {
-    // Return all products this customer has ordered
-    todoCollectionTask()
-}
+val Customer.orderedProducts: Set<Product>
+    get() {
+        // Return all products this customer has ordered
+        todoCollectionTask()
+    }
 
-val Shop.allOrderedProducts: Set<Product> get() {
-    // Return all products that were ordered by at least one customer
-    todoCollectionTask()
-}
+val Shop.allOrderedProducts: Set<Product>
+    get() {
+        // Return all products that were ordered by at least one customer
+        todoCollectionTask()
+    }
