@@ -7,6 +7,6 @@ fun example6() {
 
 fun Customer.getTotalOrderPrice(): Double {
     // Return the sum of prices of all products that a customer has ordered.
-    // Note: a customer may order the same product for several times.
+    // Note: a customer may order the same product several times.
     todoCollectionTask()
 }
