@@ -8,14 +8,13 @@ fun example9() {
 // The same as
 fun whatFoldDoes(): Int {
     var result = 1
-    listOf(1, 2, 3, 4).forEach { element -> result = element * result}
+    listOf(1, 2, 3, 4).forEach { element -> result = element * result }
     return result
 }
 
 fun Shop.getSetOfProductsOrderedByEachCustomer(): Set<Product> {
     // Return the set of products that were ordered by each of the customers
-    return customers.fold(allOrderedProducts, {
-        orderedByAll, customer ->
+    return customers.fold(allOrderedProducts, { orderedByAll, customer ->
         todoCollectionTask()
     })
 }
