@@ -1,6 +1,6 @@
 package iii_conventions
 
-import iii_conventions.TimeInterval.*
+import iii_conventions.TimeInterval.DAY
 import java.util.*
 
 fun MyDate.nextDay() = addTimeIntervals(DAY, 1)
