@@ -14,40 +14,40 @@ open class Tag(val name: String) {
     }
 }
 
-class Attribute(val name : String, val value : String) {
+class Attribute(val name: String, val value: String) {
     override fun toString() = """$name="$value""""
 }
 
-fun <T: Tag> T.set(name: String, value: String?): T {
+fun <T : Tag> T.set(name: String, value: String?): T {
     if (value != null) {
         attributes.add(Attribute(name, value))
     }
     return this
 }
 
-fun <T: Tag> Tag.doInit(tag: T, init: T.() -> Unit): T {
+fun <T : Tag> Tag.doInit(tag: T, init: T.() -> Unit): T {
     tag.init()
     children.add(tag)
     return tag
 }
 
-class Html: Tag("html")
-class Table: Tag("table")
-class Center: Tag("center")
-class TR: Tag("tr")
-class TD: Tag("td")
-class Text(val text: String): Tag("b") {
+class Html : Tag("html")
+class Table : Tag("table")
+class Center : Tag("center")
+class TR : Tag("tr")
+class TD : Tag("td")
+class Text(val text: String) : Tag("b") {
     override fun toString() = text
 }
 
 fun html(init: Html.() -> Unit): Html = Html().apply(init)
 
-fun Html.table(init : Table.() -> Unit) = doInit(Table(), init)
-fun Html.center(init : Center.() -> Unit) = doInit(Center(), init)
+fun Html.table(init: Table.() -> Unit) = doInit(Table(), init)
+fun Html.center(init: Center.() -> Unit) = doInit(Center(), init)
 
-fun Table.tr(color: String? = null, init : TR.() -> Unit) = doInit(TR(), init).set("bgcolor", color)
+fun Table.tr(color: String? = null, init: TR.() -> Unit) = doInit(TR(), init).set("bgcolor", color)
 
-fun TR.td(color: String? = null, align : String = "left", init : TD.() -> Unit) = doInit(TD(), init).set("align", align).set("bgcolor", color)
+fun TR.td(color: String? = null, align: String = "left", init: TD.() -> Unit) = doInit(TD(), init).set("align", align).set("bgcolor", color)
 
-fun Tag.text(s : Any?) = doInit(Text(s.toString()), {})
+fun Tag.text(s: Any?) = doInit(Text(s.toString()), {})
 
