@@ -8,7 +8,7 @@ import javax.swing.SwingConstants.CENTER
 
 
 fun main(args: Array<String>) {
-    with (JFrame("Product popularity")) {
+    with(JFrame("Product popularity")) {
         setSize(600, 600)
         defaultCloseOperation = JFrame.EXIT_ON_CLOSE
         add(JScrollPane(JLabel(renderProductTable(), CENTER)))
